Check how many times a variable is forwarded.
In MSL there's a hard limit of 256 parenthesis, and this happens when
the variable is forwarded that number of times, so we check if the
operation is the same until 255 times and then break the forwarding
chain.

Since this operation is single threaded static variables are ok.

This would fix #1760.

Signed-off-by: Sebastián Aedo <saedo@codeweavers.com>